+++ Sources/MapboxCoreNavigation/EventDetails.swift
@@ -99,10 +99,7 @@ extension GlobalEventDetails {
     var operatingSystem: String { "\(ProcessInfo.systemName) \(ProcessInfo.systemVersion)" }
     var platform: String { ProcessInfo.systemName }
     var sdkVersion: String {
-        guard let stringForShortVersion = Bundle.string(forMapboxCoreNavigationInfoDictionaryKey: "CFBundleShortVersionString") else {
-            preconditionFailure("CFBundleShortVersionString must be set in the Info.plist.")
-        }
-        return stringForShortVersion
+        return Bundle.navigationSDKVersion
     }
     var screenBrightness: Int { Int(UIScreen.main.brightness * 100) }
     var volumeLevel: Int { Int(AVAudioSession.sharedInstance().outputVolume * 100) }